break up main entry point into two modules, one for querying and one for field processing
